finished file restructurer
