+++ client/src/controllers/Role.js
@@ -24,6 +24,10 @@ class Role {
         this.behavior = params?.behavior;
     }
 
+    order = () => {
+
+    }
+
 }
 
 export { Role };
