+++ client/src/views/GameSettings.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import ErrorModal from './components/ErrorModal';
-import Connection from '../workers/Conncetion';
+import Connection from '../controllers/Conncetion';
 import Behavior from './Behavior';
 import BehaviorsList from './components/BehaviorsList';
 
-import { defaultBehavior } from '../workers/Behaviors';
+import { defaultBehavior } from '../controllers/Behaviors';
 
 const GameSettings = ({ }) => {
   // State for the game settings
